moved state to state.ts
